requirements.txt
env== 0.1.0
openai==0.27.6
python-dotenv==1.0.1
slither-analyzer==0.10.0
typer==0.9.0